backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1

# HTTP client
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.3

# Compression
brotli==1.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.2.0

# Utilities
email-validator==2.1.0
python-dateutil==2.8.2
pytz==2023.3

# PDF Export
reportlab==4.0.4
arabic-reshaper==3.0.0
python-bidi==0.4.2

# Development dependencies
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
isort==5.13.2
flake8==7.0.0